+++ deode/suites/deode_suite_components.py
@@ -82,6 +82,7 @@ class E923MonthlyFamily(EcflowSuiteFamily):
         trigger=None,
         ecf_files_remotely=None,
         dry_run: bool = False,
+        limit=None,
     ):
         """Class initialization."""
         super().__init__(
@@ -91,11 +92,6 @@ class E923MonthlyFamily(EcflowSuiteFamily):
             trigger=trigger,
             ecf_files_remotely=ecf_files_remotely,
         )
-
-        if not dry_run:
-            e923max = config["suite_control.max_monthly_tasks"]
-            e923_limit = EcflowSuiteLimit("e923_limit", e923max)
-            self.ecf_node.add_limit(e923_limit.limit_name, e923_limit.max_jobs)
 
         seasons = {
             "Q1": "01,02,03",
@@ -128,7 +124,7 @@ class E923MonthlyFamily(EcflowSuiteFamily):
                 self,
                 ecf_files,
                 ecf_files_remotely=ecf_files_remotely,
-                limit=e923_limit if not dry_run else None,
+                limit=limit if not dry_run else None,
             )
 
             EcflowSuiteTask(
@@ -156,6 +152,7 @@ class PgdNode(EcflowSuiteFamily, EcflowSuiteTask):
         ecf_files,
         trigger=None,
         ecf_files_remotely=None,
+        limit=None,
     ):
         """Class initialization."""
         if config["pgd.one_decade"]:
@@ -178,6 +175,7 @@ class PgdNode(EcflowSuiteFamily, EcflowSuiteTask):
                     self,
                     ecf_files,
                     ecf_files_remotely=ecf_files_remotely,
+                    limit=limit,
                 )
 
                 EcflowSuiteTask(
@@ -230,6 +228,13 @@ class StaticDataFamily(EcflowSuiteFamily):
             ecf_files_remotely=ecf_files_remotely,
         )
 
+        if not dry_run:
+            static_data_max = config["suite_control.max_static_data_tasks"]
+            static_data_limit = EcflowSuiteLimit("static_data_limit", static_data_max)
+            self.ecf_node.add_limit(
+                static_data_limit.limit_name, static_data_limit.max_jobs
+            )
+
         pgd_node = None
         if config["suite_control.do_pgd"]:
             pgd_input = PgdInputFamily(
@@ -273,6 +278,7 @@ class StaticDataFamily(EcflowSuiteFamily):
             trigger=e923constant,
             ecf_files_remotely=ecf_files_remotely,
             dry_run=dry_run,
+            limit=static_data_limit if not dry_run else None,
         )
 
         if config["suite_control.do_pgd"]:
@@ -285,6 +291,7 @@ class StaticDataFamily(EcflowSuiteFamily):
                 ecf_files,
                 trigger=e923constant,
                 ecf_files_remotely=ecf_files_remotely,
+                limit=static_data_limit if not dry_run else None,
             )
         else:
             pgd_update = None